(test): splitting tets for linter
